backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
